# docs/swagger.yaml
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      confirm_password:
        example: NewPass123!
        type: string
      current_password:
        example: OldPass123!
        type: string
      new_password:
        example: NewPass123!
        type: string
    required:
    - confirm_password
    - current_password
    - new_password
    type: object
...
  dto.ForgotPasswordRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
...
  dto.LoginRequest:
    properties:
      device_id:
        example: device_12345
        type: string
      email_or_username:
        example: user@example.com
        type: string
      password:
        example: SecurePass123!
        type: string
      remember_me:
        example: true
        type: boolean
    required:
    - email_or_username
    - password
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      session_id:
        example: sess_123456789
        type: string
      user:
        $ref: '#/definitions/dto.UserInfo'
    type: object
...
  dto.RefreshTokenRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      confirm_password:
        example: SecurePass123!
        type: string
      email:
        example: user@example.com
        type: string
      password:
        example: SecurePass123!
        type: string
      username:
        example: johndoe
        maxLength: 30
        minLength: 3
        type: string
    required:
    - confirm_password
    - email
    - password
    - username
    type: object
  dto.RegisterResponse:
    properties:
      message:
        example: Registration successful. Please check your email for verification.
        type: string
      requires_verification:
        example: true
        type: boolean
      user_id:
        example: usr_123456789
        type: string
    type: object
  dto.ResendVerificationRequest:
    properties:
      email:
        example: user@example.com
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordRequest:
    properties:
      code:
        example: "123456"
        type: string
      confirm_password:
        example: NewPass123!
        type: string
      new_password:
        example: NewPass123!
        type: string
    required:
    - code
    - confirm_password
    - new_password
    type: object
...
  dto.UserInfo:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      email_verified:
        example: true
        type: boolean
      id:
        example: usr_123456789
        type: string
      last_login_at:
        example: "2023-01-15T10:30:00Z"
        type: string
      role:
        example: user
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  dto.VerifyEmailRequest:
    properties:
      code:
        example: "123456"
        type: string
      email:
        example: user@example.com
        type: string
    required:
    - code
    - email
    type: object
...
  shared.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/change-password:
    post:
      consumes:
      - application/json
      description: Change password for authenticated user
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Current and new password
        in: body
        name: changeRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Change password
      tags:
      - auth
...
  /api/v1/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset email to user
      parameters:
      - description: Email address
        in: body
        name: forgotRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Request password reset
      tags:
      - auth
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return access token
      parameters:
      - description: Login credentials
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
      summary: Login user
      tags:
      - auth
  /api/v1/logout:
    post:
      consumes:
      - application/json
      description: Invalidate current session
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - auth
  /api/v1/logout-all:
    post:
      consumes:
      - application/json
      description: Invalidate all sessions for the user
      parameters:
      - default: Bearer <user_token>
        description: User Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - Bearer: []
      summary: Logout from all devices
      tags:
      - auth
  /api/v1/refresh:
    post:
      consumes:
      - application/json
      description: Generate new access token using refresh token
      parameters:
      - description: Refresh token
        in: body
        name: refreshRequest
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email verification and password
        confirmation
      parameters:
      - description: Registration details with password confirmation
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/resend-verification:
    post:
      consumes:
      - application/json
      description: Send a new verification email to user
      parameters:
      - description: Email address
        in: body
        name: resendRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Resend verification email
      tags:
      - auth
  /api/v1/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password using reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: resetRequest
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Reset password
      tags:
      - auth
...
  /api/v1/username/check/{username}:
    get:
      consumes:
      - application/json
      description: Check if username is available for registration
      parameters:
      - description: Username to check
        in: path
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
      summary: Check username availability
      tags:
      - auth
  /api/v1/verify-email:
    post:
      consumes:
      - application/json
      description: Verify user email with 6-digit verification code
      parameters:
      - description: Verification code and email
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/shared.Response'
            - properties:
                data:
                  type: object
              type: object
      summary: Verify email
      tags:
      - auth
swagger: "2.0"